moving exported function into utils; move tests into utils.test.ts
